Refactored code to support git history.| #1253
